Consolidate sources using scala-collection-compat
