+++ core/src/main/scala/com/banno/kafka/admin/AdminImpl.scala
@@ -16,6 +16,7 @@
 
 package com.banno.kafka.admin
 
+import scala.collection.compat._
 import cats.effect.Sync
 import org.apache.kafka.common.TopicPartitionReplica
 import org.apache.kafka.common.config.ConfigResource
