+++ core/src/main/scala/com/banno/kafka/consumer/ConsumerApi.scala
@@ -16,6 +16,7 @@
 
 package com.banno.kafka.consumer
 
+import scala.collection.compat._
 import fs2.Stream
 import cats.effect._
 import cats.implicits._
