+++ core/src/main/scala/com/banno/kafka/consumer/ConsumerOps.scala
@@ -16,6 +16,7 @@
 
 package com.banno.kafka.consumer
 
+import scala.collection.compat._
 import cats._
 import cats.implicits._
 import fs2._
