+++ core/src/main/scala/com/banno/kafka/metrics/IOMetricsReporter.scala
@@ -16,6 +16,7 @@
 
 package com.banno.kafka.metrics
 
+import scala.collection.compat._
 import org.apache.kafka.common.metrics.{KafkaMetric, MetricsReporter}
 import cats.effect.IO
 import java.util.{Map => JMap, List => JList}
