+++ core/src/main/scala/com/banno/kafka/producer/ProducerApi.scala
@@ -16,6 +16,7 @@
 
 package com.banno.kafka.producer
 
+import scala.collection.compat._
 import cats.implicits._
 import cats.effect.{Async, Resource, Sync}
 import java.util.concurrent.{Future => JFuture}
