+++ core/src/main/scala/com/banno/kafka/schemaregistry/SchemaRegistryApi.scala
@@ -16,6 +16,7 @@
 
 package com.banno.kafka.schemaregistry
 
+import scala.collection.compat._
 import org.apache.avro.Schema
 import io.confluent.kafka.schemaregistry.client.{CachedSchemaRegistryClient, SchemaMetadata}
 import io.confluent.kafka.schemaregistry.client.rest.RestService
