+++ core/src/main/scala/com/banno/kafka/schemaregistry/SchemaRegistryOps.scala
@@ -16,6 +16,7 @@
 
 package com.banno.kafka.schemaregistry
 
+import scala.collection.compat._
 import org.apache.avro.Schema
 import com.sksamuel.avro4s.{DefaultFieldMapper, SchemaFor}
 import cats.FlatMap
